Seperate block label and feature label layouts.
