+++ src/canvas/geometry/FeatureLabel.js
@@ -16,21 +16,32 @@ export class FeatureLabel extends SceneGraphNodeBase {
    * @param parent - parent scene graph node
    * @param bioMap - map data
    * @param featureModel - feature data
+   * @param config - base configuration
    */
 
   constructor({parent, bioMap, featureModel,config}) {
     super({parent, tags: [featureModel.name]});
     this.config = config;
-    this.model = featureModel;
+    this.labelPos = config.labelPosition || config.position;
+    this.fm = featureModel;
+    this.model = featureModel.model;
     this.view = bioMap.view;
     this.pixelScaleFactor = this.view.pixelScaleFactor;
     this.invert = bioMap.view.invert;
-    this.start = this.model.coordinates.start;
+    this.position = this.invert ? this.model.coordinates.start : this.model.coordinates.stop;
+    let y1 = translateScale(this.position, this.view.base, this.view.visible, this.invert) * this.pixelScaleFactor;
+    this.show = true;
+    let padding = 2;
+    let width = config.labelSize + padding ;
+    this.offset = this.labelPos >= 0 ? 0 : padding;
+    let left = this.labelPos >= 0 ? this.fm.bounds.right  : this.fm.bounds.left-width;
+    this.show = false;
+
     this.bounds = new Bounds({
-      top: 0,
-      left: 5,
-      width: 200, //this.fontSize*(this.model.name.length),
-      height: 12,
+      top: y1,
+      left: left,
+      width: width,
+      height: -this.config.labelSize * this.model.name.length/2,
       allowSubpixel: false
     });
   }
@@ -42,16 +53,20 @@ export class FeatureLabel extends SceneGraphNodeBase {
 
   draw(ctx) {
     let config = this.config;
-    let y = translateScale(this.start, this.view.base, this.view.visible, this.invert) * this.pixelScaleFactor;
-    this.bounds.top = y;
-    this.bounds.bottom = y + config.labelSize;
-    let gb = this.globalBounds || {};
     ctx.font = `${config.labelSize}px ${config.labelFace}`;
+    let y1 = translateScale(this.position, this.view.base, this.view.visible, this.invert) * this.pixelScaleFactor;
+    this.bounds.top = y1;
+    this.bounds.height = -ctx.measureText(this.model.name).width;
+    if(!this.show) return;
+    let gb = this.globalBounds || {};
+    ctx.save();
+    ctx.translate(gb.right, gb.top);
     ctx.textAlign = 'left';
     ctx.fillStyle = config.labelColor;
-    ctx.fillText(this.model.name, gb.left, gb.top);
+    ctx.rotate(-Math.PI / 2);
+    ctx.fillText(this.model.name, 0, -this.offset); // config.labelSize-2);
+    ctx.restore();
+    this.show = false; // reset show for next draw cycle
     // reset bounding box to fit the new stroke location/width
-    this.bounds.width = this.bounds.left + Math.floor(ctx.measureText(this.model.name).width) + 1;
-    if (this.parent.bounds.width < this.bounds.width) this.parent.bounds.width = this.bounds.width;
-  }
+     }
 }
